requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
docker
six
jinja2
gitpython